requirements.txt
python-telegram-bot==20.3
python-dotenv==1.1.0
spacy[en_core_web_sm]==3.6.0
dateparser==1.1.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.30.2
torch==2.0.1
huggingface-hub==0.19.4
python-magic==0.4.27
pyinstaller==6.2.0
tensorboard==2.19.0
numpy==1.26.0 